Day 2 Afternoon Update
Decided to basically be a TA for everyone else for about 6 hours
straight, meaning I didn't work on my own sites.  Made an ultimatum
preventing that from now on, but I'm exhausted from two nights in a row
with 4 hours of sleep.  I think I understand how to strategize and
format HTML correctly, and I'll follow up on that tomorrow to make sure
of that.